refactor(sync): extract RefreshDatabase into refresh.go (#83 item 4)
Pure code move, no behavior change. database.go was 651 lines with the
~240-line RefreshDatabase function dominating it.

Moved to internal/sync/refresh.go:
- RefreshOptions
- RefreshDatabase
- findSubSourceFolders, refreshMultiSource (refresh-only helpers)

Left in database.go (shared or migration concerns, not refresh-specific):
- timestampsEqual (also used by page.go, push.go, validation.go)
- scanLocalFiles / localFileInfo / markAsDeleted (shared fs helpers)
- removeLegacySQLite / migrateToUUIDFilenames (one-off migrations — issue
  #83 item 3 territory, deliberately untouched)

Verified: removed lines from database.go are byte-identical to refresh.go's
body; go build, go vet, go test ./... all pass.

Related to #83
